Make codesandbox edit work again
