+++ docs/src/pages/components/modal/SimpleModal.tsx
@@ -23,6 +23,7 @@ const useStyles = makeStyles((theme: Theme) =>
       position: 'absolute',
       width: 400,
       backgroundColor: theme.palette.background.paper,
+      border: '2px solid #000',
       boxShadow: theme.shadows[5],
       padding: theme.spacing(2, 4, 4),
       outline: 'none',
