+++ docs/src/pages/customization/palette/palette.md
@@ -176,4 +176,4 @@ const theme = createMuiTheme({
 });
 ```
 
-{{"demo": "pages/customization/palette/DarkTheme.js", "hideEditButton": true}}
+{{"demo": "pages/customization/palette/DarkTheme.js"}}
